+++ src/Aptabase.Core/EventData.cs
@@ -1,0 +1,17 @@
+﻿namespace Aptabase.Core;
+
+internal class EventData
+{
+    public string Timestamp { get; set; }
+    public string EventName { get; set; }
+    public Dictionary<string, object>? Props { get; set; }
+    public SystemInfo? SystemProps { get; set; }
+    public string? SessionId { get; set; }
+
+    public EventData(string eventName, Dictionary<string, object>? props = null)
+    {
+        Timestamp = DateTime.UtcNow.ToString("o");
+        EventName = eventName;
+        Props = props;
+    }
+}