+++ src/Aptabase.Core/IAptabaseClient.cs
@@ -1,0 +1,10 @@
+﻿namespace Aptabase.Core;
+
+/// <summary>
+/// Aptabase client used for tracking events
+/// </summary>
+public interface IAptabaseClient : IAsyncDisposable
+{
+    Task TrackEvent(string eventName, Dictionary<string, object>? props = null);
+}
+