+++ src/Aptabase.Maui/MauiExtensions.cs
@@ -1,0 +1,48 @@
+﻿using Aptabase.Core;
+using Aptabase.Maui;
+using Microsoft.Extensions.Logging;
+
+namespace Microsoft.Maui.Hosting;
+
+/// <summary>
+/// Aptabase extensions for <see cref="MauiAppBuilder"/>.
+/// </summary>
+public static class MauiExtensions
+{
+    /// <summary>
+    /// Uses Aptabase integration.
+    /// </summary>
+    /// <param name="builder">The builder.</param>
+    /// <param name="appKey">The App Key.</param>
+    /// <param name="options">Initialization Options.</param>
+    /// <returns>The <paramref name="builder"/>.</returns>
+    public static MauiAppBuilder UseAptabase(this MauiAppBuilder builder, string appKey, AptabaseOptions? options = null)
+    {
+        builder.Services.AddSingleton<IAptabaseClient>(serviceProvider =>
+        {
+            IAptabaseClient client;
+            var loggerFactory = serviceProvider.GetRequiredService<ILoggerFactory>();
+
+            if (options?.EnablePersistence != true)
+            {
+                client = new AptabaseClient(appKey, options, loggerFactory.CreateLogger<AptabaseClient>());
+            }
+            else
+            {
+                client = new AptabasePersistentClient(appKey, options, loggerFactory.CreateLogger<AptabasePersistentClient>());
+            }
+
+            if (options?.EnableCrashReporting == true)
+
+            {
+                _ = new AptabaseMauiCrashReporter(client, loggerFactory.CreateLogger<AptabaseMauiCrashReporter>());
+            }
+
+            return client;
+        });
+
+        return builder;
+    }
+}
+
+