+++ test/HelloMaui/App.xaml.cs
@@ -1,0 +1,16 @@
+﻿using Aptabase.Core;
+
+namespace HelloMaui;
+
+public partial class App : Application
+{
+	public App(IAptabaseClient aptabase)
+	{
+		InitializeComponent();
+
+		MainPage = new AppShell();
+
+		aptabase.TrackEvent("app_started");
+	}
+}
+