+++ test/HelloMaui/AppShell.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace HelloMaui;
+
+public partial class AppShell : Shell
+{
+	public AppShell()
+	{
+		InitializeComponent();
+	}
+}
+